+++ packages/controls/src/controls/index.ts
@@ -11,12 +11,11 @@ export * from './select'
 export * from './shape'
 export * from './slot'
 export * from './style'
-export * from './style-v2'
 export * from './text-area'
 export * from './text-input'
 export * from './font'
 export * from './typography'
-export * from './shape-v2'
+export * from './shape'
 
 export * from './associated-types'
 export * from './definition'
