+++ packages/controls/src/controls/shape/v1/shape.test.ts
@@ -1,18 +1,18 @@
-import { createReplacementContext } from '../../context'
-import { Targets } from '../../introspection'
-import { deserializeRecord, type DeserializedRecord } from '../../serialization'
+import { createReplacementContext } from '../../../context'
+import { Targets } from '../../../introspection'
+import { deserializeRecord, type DeserializedRecord } from '../../../serialization'
 
-import { type DataType, type ValueType } from '../associated-types'
-import { Checkbox, CheckboxDefinition } from '../checkbox'
-import { Color, ColorDefinition } from '../color'
-import { Combobox } from '../combobox'
-import { ControlDefinition } from '../definition'
-import { Image } from '../image'
-import { GenericLink as Link } from '../link'
-import { List } from '../list'
-import { Number } from '../number'
-import { TextArea } from '../text-area'
-import { TextInput } from '../text-input'
+import { type DataType, type ValueType } from '../../associated-types'
+import { Checkbox, CheckboxDefinition } from '../../checkbox'
+import { Color, ColorDefinition } from '../../color'
+import { Combobox } from '../../combobox'
+import { ControlDefinition } from '../../definition'
+import { Image } from '../../image'
+import { GenericLink as Link } from '../../link'
+import { List } from '../../list'
+import { Number } from '../../number'
+import { TextArea } from '../../text-area'
+import { TextInput } from '../../text-input'
 
 import { Shape, ShapeDefinition } from './shape'
