+++ packages/controls/src/controls/shape/v1/shape.ts
@@ -1,35 +1,34 @@
 import { z } from 'zod'
 
-import { mapValues } from '../../lib/functional'
-import { StableValue } from '../../lib/stable-value'
-import { safeParse, type ParseResult } from '../../lib/zod'
+import { mapValues } from '../../../lib/functional'
+import { StableValue } from '../../../lib/stable-value'
+import { safeParse, type ParseResult } from '../../../lib/zod'
 
-import { type Data } from '../../common'
+import { type Data } from '../../../common'
 import {
   type CopyContext,
   type MergeTranslatableDataContext,
-} from '../../context'
-import { type IntrospectionTarget } from '../../introspection'
-import { type ResourceResolver } from '../../resources/resolver'
+} from '../../../context'
+import { type IntrospectionTarget } from '../../../introspection'
+import { type ResourceResolver } from '../../../resources/resolver'
 import {
   SerializationSchema,
   type DeserializedRecord,
   type SerializedRecord,
-} from '../../serialization'
-import { type Stylesheet } from '../../stylesheet'
-
+} from '../../../serialization'
+import { type Stylesheet } from '../../../stylesheet'
 import {
   type DataType as DataType_,
   type ResolvedValueType as ResolvedValueType_,
   type ValueType as ValueType_,
-} from '../associated-types'
+} from '../../associated-types'
 import {
   ControlDefinition,
   serialize,
   type Resolvable,
   type SchemaType,
-} from '../definition'
-import { type SendMessage } from '../instance'
+} from '../../definition'
+import { type SendMessage } from '../../instance'
 
 import { ShapeControl } from './shape-control'
