+++ packages/controls/src/controls/shape/v2/shape-v2-control.ts
@@ -1,9 +1,9 @@
-import { ControlDefinition } from '../definition'
+import { ControlDefinition } from '../../definition'
 import {
   ControlInstance,
   type ControlMessage,
   type SendMessage,
-} from '../instance'
+} from '../../instance'
 
 import { ShapeV2Definition } from './shape-v2'
